+++ backend/api/repository/repository_posts.go
@@ -6,7 +6,7 @@ import "escapead/projectAD/backend/api/models"
 type PostsRepository interface {
 	Save(models.Post) (models.Post, error)
 	FindAll() ([]models.Post, error)
-	FindByID(uint32) (models.Post, error)
-	Update(uint32, models.Post) (int64, error)
-	Delete(uint32) (int64, error)
+	FindByID(uint64) (models.Post, error)
+	Update(uint64, models.Post) (int64, error)
+	Delete(uint64) (int64, error)
 }
